test(select): migrate item
