+++ core/src/components/select/test/item/select.e2e.ts-snapshots/select-inset-list-no-fill-ios-ltr-Mobile-Safari-linux.png
