+++ core/src/components/select/test/item/select.e2e.ts-snapshots/select-inset-list-no-fill-md-ltr-Mobile-Firefox-linux.png
