+++ core/src/components/select/test/item/select.e2e.ts-snapshots/select-list-no-fill-ios-ltr-Mobile-Firefox-linux.png
