+++ core/src/components/select/test/item/select.e2e.ts-snapshots/select-list-no-fill-md-ltr-Mobile-Safari-linux.png
